Remove doublylinkedlist, Go already has one
Go's list is different in that it doesn't enforce polymorphic
collections, but that's not a _real_ problem IMO, was fun to code
though.
It also does not support index-based access, but that's more of a good
thing, since a double linked list isn't very suitable for that and in
the exceptionally case that index-based access should be needed, it can
be coded up quite quickly; Go's list supports a nice way for iterating
over it.